Verifing a column before dropping it

+++ db/migrate/20201222004459_remove_columns_from_county_result.rb
@@ -1,6 +1,10 @@
 class RemoveColumnsFromCountyResult < ActiveRecord::Migration[5.2]
   def change
-    remove_column :county_results, :bmp_id, :integer
-    remove_column :county_results, :crop_ids, :string
+  	if !(ActiveRecord::Base.connection.column_exists?(:county_results, :bmp_id))
+    	remove_column :county_results, :bmp_id, :integer
+    end
+    if !(ActiveRecord::Base.connection.column_exists?(:county_results, :crop_ids))
+    	remove_column :county_results, :crop_ids, :string
+    end
   end
 end
